move key cmd tests to own module
